validation changes moved from applet to HAL

+++ Applet/AndroidSEProviderLib/src/com/android/javacard/seprovider/KMEcdsa256NoDigestSignature.java
@@ -15,8 +15,8 @@
  */
 package com.android.javacard.seprovider;
 
-import javacard.security.CryptoException;
 import javacard.framework.Util;
+import javacard.security.CryptoException;
 import javacard.security.Key;
 import javacard.security.MessageDigest;
 import javacard.security.Signature;
@@ -25,7 +25,7 @@ import javacardx.crypto.Cipher;
 public class KMEcdsa256NoDigestSignature extends Signature {
 
   public static final byte ALG_ECDSA_NODIGEST = (byte) 0x67;
-  public static final short MAX_NO_DIGEST_MSG_LEN = 32;
+  public static final byte MAX_NO_DIGEST_MSG_LEN = 32;
   private byte algorithm;
   private Signature inst;
 
@@ -40,16 +40,13 @@ public class KMEcdsa256NoDigestSignature extends Signature {
   }
 
   @Override
-  public void init(Key key, byte b, byte[] bytes, short i, short i1)
-      throws CryptoException {
+  public void init(Key key, byte b, byte[] bytes, short i, short i1) throws CryptoException {
     inst.init(key, b, bytes, i, i1);
   }
 
   @Override
-  public void setInitialDigest(byte[] bytes, short i, short i1, byte[] bytes1,
-      short i2, short i3) throws CryptoException {
-
-  }
+  public void setInitialDigest(byte[] bytes, short i, short i1, byte[] bytes1, short i2, short i3)
+      throws CryptoException {}
 
   @Override
   public byte getAlgorithm() {
@@ -77,8 +74,7 @@ public class KMEcdsa256NoDigestSignature extends Signature {
   }
 
   @Override
-  public void update(byte[] message, short msgStart, short messageLength)
-      throws CryptoException {
+  public void update(byte[] message, short msgStart, short messageLength) throws CryptoException {
     // HAL accumulates the data and send it at finish operation.
   }
 
@@ -91,36 +87,46 @@ public class KMEcdsa256NoDigestSignature extends Signature {
       }
       // add zeros to the left
       if (i1 < MAX_NO_DIGEST_MSG_LEN) {
-        Util.arrayFillNonAtomic(KMAndroidSEProvider.getInstance().tmpArray,
-            (short) 0, (short) MAX_NO_DIGEST_MSG_LEN, (byte) 0);
+        Util.arrayFillNonAtomic(
+            KMAndroidSEProvider.getInstance().tmpArray,
+            (short) 0,
+            (short) MAX_NO_DIGEST_MSG_LEN,
+            (byte) 0);
       }
-      Util.arrayCopyNonAtomic(bytes, i,
+      Util.arrayCopyNonAtomic(
+          bytes,
+          i,
           KMAndroidSEProvider.getInstance().tmpArray,
-          (short) (MAX_NO_DIGEST_MSG_LEN - i1), i1);
-      return inst.signPreComputedHash(KMAndroidSEProvider.getInstance().tmpArray,
-          (short) 0, (short) MAX_NO_DIGEST_MSG_LEN, bytes1, i2);
+          (short) (MAX_NO_DIGEST_MSG_LEN - i1),
+          i1);
+      return inst.signPreComputedHash(
+          KMAndroidSEProvider.getInstance().tmpArray,
+          (short) 0,
+          (short) MAX_NO_DIGEST_MSG_LEN,
+          bytes1,
+          i2);
     } finally {
       KMAndroidSEProvider.getInstance().clean();
     }
   }
 
   @Override
-  public short signPreComputedHash(byte[] bytes, short i, short i1,
-      byte[] bytes1, short i2) throws CryptoException {
+  public short signPreComputedHash(byte[] bytes, short i, short i1, byte[] bytes1, short i2)
+      throws CryptoException {
     return inst.sign(bytes, i, i1, bytes1, i2);
   }
 
   @Override
-  public boolean verify(byte[] bytes, short i, short i1, byte[] bytes1,
-      short i2, short i3) throws CryptoException {
-    //Verification is handled inside HAL
+  public boolean verify(byte[] bytes, short i, short i1, byte[] bytes1, short i2, short i3)
+      throws CryptoException {
+    // Verification is handled inside HAL
     return false;
   }
 
   @Override
-  public boolean verifyPreComputedHash(byte[] bytes, short i, short i1,
-      byte[] bytes1, short i2, short i3) throws CryptoException {
-    //Verification is handled inside HAL
+  public boolean verifyPreComputedHash(
+      byte[] bytes, short i, short i1, byte[] bytes1, short i2, short i3) throws CryptoException {
+    // Verification is handled inside HAL
     return false;
   }
 }
